Create base converter and remove help class
Also fixes. #56
