+++ src/Our.Umbraco.Ditto/ComponentModel/TypeConverters/DittoUltimatePickerConverter.cs
@@ -1,17 +1,19 @@
 ﻿namespace Our.Umbraco.Ditto
 {
     using System;
+    using System.Collections.Generic;
     using System.ComponentModel;
     using System.Diagnostics;
     using System.Globalization;
     using System.Linq;
 
+    using global::Umbraco.Core;
     using global::Umbraco.Core.Models;
 
     /// <summary>
-    /// Provides a unified way of converting content picker properties to strong typed model.
+    /// Provides a unified way of converting ultimate picker properties to strong typed collections.
     /// </summary>
-    public class ContentPickerConverter : TypeConverter
+    public class DittoUltimatePickerConverter : DittoBaseConverter
     {
         /// <summary>
         /// Returns whether this converter can convert an object of the given type to the type of this converter, using the specified context.
@@ -27,8 +29,7 @@
             // ReSharper disable once ConditionIsAlwaysTrueOrFalse
             if (sourceType == null
                 || sourceType == typeof(string)
-                || sourceType == typeof(int)
-                || typeof(IPublishedContent).IsAssignableFrom(sourceType))
+                || sourceType == typeof(int))
             {
                 return true;
             }
@@ -47,11 +48,6 @@
         /// </returns>
         public override object ConvertFrom(ITypeDescriptorContext context, CultureInfo culture, object value)
         {
-            if (value == null)
-            {
-                return null;
-            }
-
             Debug.Assert(context.PropertyDescriptor != null, "context.PropertyDescriptor != null");
             var propertyType = context.PropertyDescriptor.PropertyType;
             var isGenericType = propertyType.IsGenericType;
@@ -59,48 +55,71 @@
                                 ? propertyType.GenericTypeArguments.First()
                                 : propertyType;
 
-            // DictionaryPublishedContent 
-            IPublishedContent content = value as IPublishedContent;
-            if (content != null)
+            if (value == null)
             {
-                return content.As(targetType, null, null, culture);
-            }
+                if (isGenericType)
+                {
+                    return EnumerableInvocations.Empty(targetType);
+                }
 
-            if (value is int)
-            {
-                return this.ConvertFromInt((int)value, targetType, culture);
-            }
-
-            int id;
-            var s = value as string;
-            if (s != null && int.TryParse(s, NumberStyles.Any, culture, out id))
-            {
-                return this.ConvertFromInt(id, targetType, culture);
-            }
-
-            return base.ConvertFrom(context, culture, value);
-        }
-
-        /// <summary>
-        /// Takes a content node ID, gets the corresponding <see cref="T:Umbraco.Core.Models.IPublishedContent"/> object,
-        /// then converts the object to the desired type.
-        /// </summary>
-        /// <param name="id">The content node ID.</param>
-        /// <param name="targetType">
-        /// The property <see cref="Type"/> to convert to.</param>
-        /// <param name="culture">The <see cref="CultureInfo" /> to use as the current culture.</param>
-        /// <returns>
-        /// An <see cref="T:System.Object"/> that represents the converted value.
-        /// </returns>
-        private object ConvertFromInt(int id, Type targetType, CultureInfo culture)
-        {
-            if (id <= 0)
-            {
                 return null;
             }
 
-            var umbracoHelper = ConverterHelper.UmbracoHelper;
-            return umbracoHelper.TypedContent(id).As(targetType, null, null, culture);
+            // If a single item is selected, this comes back as an int, not a string.
+            if (value is int)
+            {
+                var id = (int)value;
+                var umbracoHelper = UmbracoHelper;
+
+                // CheckBoxList, ListBox
+                if (targetType != null)
+                {
+                    return umbracoHelper.TypedContent(id)
+                                        .As(targetType, null, null, culture).YieldSingleItem();
+                }
+
+                // AutoComplete, DropDownList, RadioButton
+                return umbracoHelper.TypedContent(id).As(propertyType, null, null, culture);
+            }
+
+            string s = value as string ?? value.ToString();
+            if (!string.IsNullOrWhiteSpace(s))
+            {
+                int n;
+                var nodeIds = s
+                    .ToDelimitedList()
+                    .Select(x => int.TryParse(x, NumberStyles.Any, culture, out n) ? n : -1)
+                    .Where(x => x > 0)
+                    .ToArray();
+
+                if (nodeIds.Any())
+                {
+                    var umbracoHelper = UmbracoHelper;
+                    var ultimatePicker = new List<IPublishedContent>();
+
+                    // ReSharper disable once LoopCanBeConvertedToQuery
+                    foreach (var nodeId in nodeIds)
+                    {
+                        var item = umbracoHelper.TypedContent(nodeId);
+
+                        if (item != null)
+                        {
+                            ultimatePicker.Add(item);
+                        }
+                    }
+
+                    // CheckBoxList, ListBox
+                    if (isGenericType)
+                    {
+                        return ultimatePicker.As(targetType, null, null, null, culture);
+                    }
+
+                    // AutoComplete, DropDownList, RadioButton
+                    return ultimatePicker.As(targetType, null, null, null, culture).FirstOrDefault();
+                }
+            }
+
+            return base.ConvertFrom(context, culture, value);
         }
     }
 }
