+++ src/Our.Umbraco.Ditto/Our.Umbraco.Ditto.csproj
@@ -83,15 +83,15 @@
     <Compile Include="Common\EnumerableInvocations.cs" />
     <Compile Include="Common\MethodBaseCacheItem.cs" />
     <Compile Include="ComponentModel\PublishedContentContext.cs" />
-    <Compile Include="ComponentModel\TypeConverters\EnumConverter.cs" />
-    <Compile Include="ComponentModel\TypeConverters\MemberPickerConverter.cs" />
-    <Compile Include="ComponentModel\TypeConverters\UltimatePickerConverter.cs" />
-    <Compile Include="ComponentModel\TypeConverters\MultiNodeTreePickerConverter.cs" />
-    <Compile Include="ComponentModel\TypeConverters\MultipleMediaPickerConverter.cs" />
-    <Compile Include="ComponentModel\TypeConverters\MediaPickerConverter.cs" />
-    <Compile Include="ComponentModel\TypeConverters\ContentPickerConverter.cs" />
-    <Compile Include="ComponentModel\ConverterHelper.cs" />
-    <Compile Include="ComponentModel\TypeConverters\HtmlStringConverter.cs" />
+    <Compile Include="ComponentModel\TypeConverters\DittoBaseConverter.cs" />
+    <Compile Include="ComponentModel\TypeConverters\DittoEnumConverter.cs" />
+    <Compile Include="ComponentModel\TypeConverters\DittoMemberPickerConverter.cs" />
+    <Compile Include="ComponentModel\TypeConverters\DittoUltimatePickerConverter.cs" />
+    <Compile Include="ComponentModel\TypeConverters\DittoMultiNodeTreePickerConverter.cs" />
+    <Compile Include="ComponentModel\TypeConverters\DittoMultipleMediaPickerConverter.cs" />
+    <Compile Include="ComponentModel\TypeConverters\DittoMediaPickerConverter.cs" />
+    <Compile Include="ComponentModel\TypeConverters\DittoContentPickerConverter.cs" />
+    <Compile Include="ComponentModel\TypeConverters\DittoHtmlStringConverter.cs" />
     <Compile Include="Extensions\EnumerableExtensions.cs" />
     <Compile Include="Extensions\TypeInferenceExtensions.cs" />
     <Compile Include="Extensions\TypeInitializationExtensions.cs" />
